validando creacion de carpeta files

+++ database/seeders/DatabaseSeeder.php
@@ -14,8 +14,12 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        Storage::deleteDirectory('files');
-        Storage::makeDirectory('files');
+        if(Storage::exists('files')){
+            Storage::deleteDirectory('files');
+            Storage::makeDirectory('files');
+        }else{
+            Storage::makeDirectory('files');
+        }
 
         // \App\Models\User::factory(10)->create();
         $this->call(PermissionSeeder::class);
